% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{append.row}
\alias{append.row}
\title{Efficiently appends row to the data.table dt.
For efficiency comparison of various methods see
http://stackoverflow.com/questions/20689650/how-to-append-rows-to-an-r-data-frame/38052208#38052208}
\usage{
append.row(dt, elems)
}
\description{
Efficiently appends row to the data.table dt.
For efficiency comparison of various methods see
http://stackoverflow.com/questions/20689650/how-to-append-rows-to-an-r-data-frame/38052208#38052208
}

