% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{count.rows.and.columns}
\alias{count.rows.and.columns}
\title{This function gets table format rendered by \code{get.table.struct} and converts it into even more table-friendly
data representation.}
\usage{
count.rows.and.columns(ans)
}
\description{
The return value is a list of max 4 elements:
* total row.count - if we keep on item on one row, this is the max of all columns in the resulting table
* common.column.count - either 1 or 0, depending on whether the common item column is present
* common.items - item numbers of all items that go to the common group (if one exists)
* discrete.column.count - total number of all columns that contain exclusive groups. More than one, if the
  search was recursive.
* discrete.items - list where each element describes every distinct exclusive group. The data format
  is the same as the main object (this is a recursively defined list)
}

