% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{get.excel.table}
\alias{get.excel.table}
\title{This is the final function that outputs the table in an Excel-friendly format.}
\usage{
get.excel.table(car, legend = NULL, tab = NULL, coloffset = 1,
  rowoffset = 1, nextid = 1)
}
\description{
All cells that are marked with the same non-NA symbol, should be merged into one group.
}

