% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{get.table.struct}
\alias{get.table.struct}
\title{Function that renders the the row into a more table-friendly struct, that can get relatively easly printed}
\usage{
get.table.struct(row)
}
\value{
A list with two elements
\describe{
 \item{\code{common.items}}{character vector with values that will appear in rows
 on the left column in the result table
 }
 \item{\code{discrete.items}}{list with one item for each element on the right column(s).
 Each item is a list with two elements: \code{common.items} and \code{discrete.items} all described above
 }
}
}
\description{
Function that renders the the row into a more table-friendly struct, that can get relatively easly printed
}

