% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostHocGroups.R
\name{GroupPostHocs.dist}
\alias{GroupPostHocs.dist}
\title{Groups elements of the given \code{dissimilarity matrix}}
\usage{
\method{GroupPostHocs}{dist}(dissimiliarity.matrix, means = NULL,
  max.recursive.level = 1, solutions.count = 1)
}
\arguments{
\item{dissimiliarity.matrix}{Object of \code{dist} class.}

\item{means}{Optional parameter. Requires numeric vector of the size of number of columns of
the dissimiliarity.matrix or \code{NULL} (default). If specified, function will sort
the dissimiliarity.matrix according to those values during the pre-process stage of the
algorithm.}

\item{max.recursive.level}{Defaults to 1 for flat representation.
If more, algorithm will try to recursively subpartition groups that have
more than 2 elements.}

\item{solutions.count}{Number of sultions to report. Defaults to 1, i.e. only the best solution
will be printed.}
}
\value{
Object of class \code{PostHocGroups}. If only one possible solution was requested,
the object has also class \code{PostHocGroup}.
}
\description{
This is the most general way to invoke the algorithm. User needs to provide distance matrix
that is interpreted as dissimilarity matrix, with value that is interpreted as a
measure of dissimilarity between items:
\describe{
 \item{matrix value >0}{Two items are similar}
 \item{matrix value <0}{Two items are different}
}
}
\details{
To get meaningfull results, user should either provide matrix with items (columns/rows)
of the matrix are sorted according to the corresponding value of the tested item, or
provide the \code{means} argument and the algorithm will sort the elements internally.
}

