% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostHocGroups.R
\name{GroupPostHocs.glht}
\alias{GroupPostHocs.glht}
\title{Groups result from the \code{glht} post-hoc test}
\usage{
\method{GroupPostHocs}{glht}(res, max.recursive.level = 1,
  solutions.count = 1)
}
\arguments{
\item{res}{Object of \code{glht} class.}

\item{max.recursive.level}{Defaults to 1 for flat representation.
If more, algorithm will try to recursively subpartition groups that have
more than 2 elements.}

\item{solutions.count}{Number of sultions to report. Defaults to 1, i.e. only the best solution
will be printed.}
}
\value{
Object of class \code{PostHocGroups}. If only one possible solution was requested,
the object has also class \code{PostHocGroup}.
}
\description{
This is a specialization of the generic function \code{GroupPostHocs}
}

