% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostHocGroups.R
\name{GroupPostHocs}
\alias{GroupPostHocs}
\title{Groups result from post-hoc test}
\usage{
GroupPostHocs(object, ...)
}
\arguments{
\item{object}{Object of one of the supported types.}

\item{max.recursive.level}{Defaults to 1 for flat representation.
If more, algorithm will try to recursively subpartition groups that have
more than 2 elements.}

\item{solutions.count}{Number of sultions to report. Defaults to 1, i.e. only the best solution
will be printed.}
}
\value{
Object of class \code{PostHocGroups}. If only one possible solution was requested,
the object has also class \code{PostHocGroup}.
}
\description{
The function is generic and can consume either dissimilarity matrix (of class \code{dist},
e.g. produced by \code{stats::dist} function), or
\code{glht} object, produced by the \code{\link[multcomp]{glht}} function
from the \code{multcomp} package.
}

