% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{makeprox}
\alias{makeprox}
\title{Creates dissimilarity matrix out from row format got from hierarchical clustering}
\usage{
makeprox(mycutree)
}
\description{
Creates dissimilarity matrix out from row format got from hierarchical clustering
}

