% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PostHocGroups.R
\name{recursive.clustering}
\alias{recursive.clustering}
\title{This is the main function.}
\usage{
recursive.clustering(myprox, prefix, solutions, recursive.level,
  max.recursive.level, means = NULL, flag.smile.correct = TRUE,
  flag.find.common.set = TRUE)
}
\description{
This is the main function.
}

