% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{symmetric.list.outer}
\alias{symmetric.list.outer}
\title{Returns symmetric simmilarity matrix based on function that returns whether a pair is significant}
\usage{
symmetric.list.outer(items, fun, ...)
}
\arguments{
\item{items}{List with items the \code{fun} operates on.}

\item{fn}{Function with signature \code{fun(item1, item2, ...)} that returns 1 or 0}
}
\value{
Rectangular symmetric matrix with return values from function \code{fn}
}
\description{
Returns symmetric simmilarity matrix based on function that returns whether a pair is significant
}

